first check in for my-utility-project This started as a little hobby project and slowly started to evolve
It uses Spring boot, hibernate and links to a postgres database.
the database gets pulled as a docker image and configured.

Liquibase scrpt runs to create the schema for the entities to map to.

Hibernate is the jpa provider being used.

The disastrous bit for now is incorporating a coreui.io template into thymeleef
and making sure the resources are correctly linked. Expect to see more changes here.

The application is still a work in progress